requirements.txt
office365==0.3.15
Office365-REST-Python-Client==2.5.14
python-dotenv==1.0.1
